requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
fastapi==0.70.0
psycopg2==2.8.5
uvicorn==0.15.0
python-decouple==3.5.0